feat: 카테고리 전체 조회
