+++ src/main/java/com/greedy/zupzup/alert/KeywordAlert.java
@@ -2,7 +2,7 @@ package com.greedy.zupzup.alert;
 
 import com.greedy.zupzup.global.BaseTimeEntity;
 import com.greedy.zupzup.member.Member;
-import com.greedy.zupzup.category.Category;
+import com.greedy.zupzup.category.domain.Category;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
