+++ src/main/java/com/greedy/zupzup/category/presentation/dto/CategoriesResponse.java
@@ -1,0 +1,6 @@
+package com.greedy.zupzup.category.presentation.dto;
+
+import com.greedy.zupzup.category.application.dto.CategoryDto;
+import java.util.List;
+
+public record CategoriesResponse(List<CategoryDto> categories) {}
