+++ src/main/java/com/greedy/zupzup/category/repository/CategoryRepository.java
@@ -1,0 +1,9 @@
+package com.greedy.zupzup.category.repository;
+
+import com.greedy.zupzup.category.domain.Category;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+    List<Category> findAllByOrderByNameAsc();
+}
